ai-gateway/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
anthropic>=0.45.0
python-dotenv>=1.0.1
# Pinning httpx to avoid the breaking changes in the SDK's current wrappers
httpx==0.27.2